Refactor sam module
